pyproject.toml
[project]
name = "nodex"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.26.0",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0",
    "bs4>=0.0.2",
    "requests>=2.32.4",
    "pytest>=8.4.1",
    "ijson>=3.4.0",
    "orjson>=3.11.1",
    "httpx[http2]>=0.28.1",
]
